requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
amadeus==8.0.0
google-search-results==2.4.2
flask==3.0.0
flask-cors==4.0.0
